Moved auth provider from domain package
