+++ fahrschein-example/src/main/java/org/zalando/fahrschein/example/ConsumerExample.java
@@ -9,7 +9,7 @@ import org.zalando.fahrschein.ExponentialBackoffStrategy;
 import org.zalando.fahrschein.Listener;
 import org.zalando.fahrschein.NakadiClient;
 import org.zalando.fahrschein.StreamParameters;
-import org.zalando.fahrschein.domain.PlatformAccessTokenProvider;
+import org.zalando.fahrschein.PlatformAccessTokenProvider;
 import org.zalando.fahrschein.domain.Subscription;
 import org.zalando.fahrschein.example.domain.MultiEventProcessor;
 import org.zalando.fahrschein.example.domain.OrderEvent;
